// src/core/socials/types/socials.ts
/* eslint-disable @typescript-eslint/naming-convention */

export interface ISocials {
	id: number;
	freebase_mid: string;
	freebase_id: string;
	imdb_id: string;
	tvrage_id: number;
	wikidata_id: string;
	facebook_id: string;
	instagram_id: string;
	tiktok_id: string;
	twitter_id: string;
	youtube_id: string;
}